refactor: use generic usefetchdata hook (#294)
