+++ app/views/ComponentAtlasView/hooks/useEditComponentAtlasForm.ts
@@ -1,11 +1,11 @@
 import { HCAAtlasTrackerComponentAtlas } from "../../../apis/catalog/hca-atlas-tracker/common/entities";
 import { PathParameter } from "../../../common/entities";
-import { useFetchComponentAtlas } from "../../../hooks/useFetchComponentAtlas";
 import { FormMethod } from "../../../hooks/useForm/common/entities";
 import { useForm } from "../../../hooks/useForm/useForm";
 import { FIELD_NAME } from "../common/constants";
 import { ComponentAtlasEditData } from "../common/entities";
 import { componentAtlasEditSchema } from "../common/schema";
+import { useFetchComponentAtlas } from "./useFetchComponentAtlas";
 
 const SCHEMA = componentAtlasEditSchema;
